Push through the 'simplest' changes
