+++ crates/bevy_ecs/src/system/mod.rs
@@ -14,7 +14,3 @@ pub use query::*;
 pub use system::*;
 pub use system_chaining::*;
 pub use system_param::*;
-
-// Electric boogaloo
-mod into_system_2;
-mod system_param_2;
